create debug.go for code shared between debug runners. Contains productFilterString for now
